Fixed the rule compile issue.
